requirements.txt
ultralytics>=8.3.0
torch>=2.2.0
opencv-python-headless>=4.10.0
albumentations>=1.4.0
numpy>=1.26.0
PyYAML>=6.0.1
pydantic>=2.7.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
streamlit>=1.37.0
onnx>=1.16.0

